+++ imhotep-server/src/main/native/optimized_ftgs/optimized/src/tgs.c
@@ -124,7 +124,7 @@ int tgs_execute_pass(struct worker_desc *worker,
     }
 
     struct bit_tree* non_zero_rows    = unpacked_table_get_non_zero_rows(group_stats);;
-    const size_t     n_rows           = unpacked_table_get_rows(group_stats);
+    const int     n_rows           = unpacked_table_get_rows(group_stats);
     uint32_t*        groups           = calloc(n_rows, sizeof(uint32_t));
     const int32_t    term_group_count = bit_tree_dump(non_zero_rows, groups, n_rows);
     int result = write_term_group_stats(session, desc, groups, term_group_count);
